refactor: centralize state clearing and move business logic to project store
